app/requirements.txt
asttokens==2.4.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
falcon==3.1.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
gunicorn==22.0.0
inflection==0.5.1
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
mariadb==1.1.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.46
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyYAML==6.0.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.1
wcwidth==0.2.13
Werkzeug==3.0.3
yml==0.0.1
